move EpochPanel to its own file
